put get_model_from_config into helpers
