+++ lightly/cli/_helpers.py
@@ -8,9 +8,14 @@ import warnings
 
 import torch
 from hydra import utils
+from torch import nn as nn
 from torch.utils.hipify.hipify_python import bcolors
 
-from lightly.models import ZOO as model_zoo
+from lightly.cli._cli_simclr import _SimCLR
+from lightly.embedding import SelfSupervisedEmbedding
+
+from lightly.models import ZOO as model_zoo, ResNetGenerator
+from lightly.models.batchnorm import get_norm_layer
 
 
 def _custom_formatwarning(msg, *args, **kwargs):
@@ -198,3 +203,44 @@ def load_from_state_dict(model,
 
     # step 3: load from checkpoint
     model.load_state_dict(state_dict, strict=strict)
+
+
+def get_model_from_config(cfg, is_cli_call: bool = False) -> SelfSupervisedEmbedding:
+    checkpoint = cfg['checkpoint']
+    if torch.cuda.is_available():
+        device = torch.device('cuda')
+    else:
+        device = torch.device('cpu')
+
+    if not checkpoint:
+        checkpoint, key = get_ptmodel_from_config(cfg['model'])
+        if not checkpoint:
+            msg = 'Cannot download checkpoint for key {} '.format(key)
+            msg += 'because it does not exist!'
+            raise RuntimeError(msg)
+        state_dict = load_state_dict_from_url(checkpoint, map_location=device)[
+            'state_dict'
+        ]
+    else:
+        checkpoint = fix_input_path(checkpoint) if is_cli_call else checkpoint
+        state_dict = torch.load(checkpoint, map_location=device)['state_dict']
+
+    # load model
+    resnet = ResNetGenerator(cfg['model']['name'], cfg['model']['width'])
+    last_conv_channels = list(resnet.children())[-1].in_features
+    features = nn.Sequential(
+        get_norm_layer(3, 0),
+        *list(resnet.children())[:-1],
+        nn.Conv2d(last_conv_channels, cfg['model']['num_ftrs'], 1),
+        nn.AdaptiveAvgPool2d(1),
+    )
+
+    model = _SimCLR(
+        features, num_ftrs=cfg['model']['num_ftrs'], out_dim=cfg['model']['out_dim']
+    ).to(device)
+
+    if state_dict is not None:
+        load_from_state_dict(model, state_dict)
+
+    encoder = SelfSupervisedEmbedding(model, None, None, None)
+    return encoder
